requirements.txt
# Core
verl>=0.2.0
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
pandas>=2.0.0
pyarrow>=14.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Training infrastructure
ray>=2.10.0
vllm>=0.4.0

# Logging
wandb>=0.16.0

# Evaluation
sympy>=1.12
numpy>=1.24.0
